Split app model boundaries for quality
