+++ Tests/QuillCodeParityTests/ParityBrowserGateTests.swift
@@ -4,17 +4,23 @@ final class ParityBrowserGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesBrowserSurfaceTypes() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let workspaceSurfaceText = try Self.appSourceText(named: "WorkspaceSurface.swift")
+        let browserStateText = try Self.appSourceText(named: "QuillCodeBrowserState.swift")
+        let browserTabStateText = try Self.appSourceText(named: "QuillCodeBrowserTabState.swift")
+        let browserSnapshotStateText = try Self.appSourceText(named: "QuillCodeBrowserSnapshotState.swift")
+        let browserCommentStateText = try Self.appSourceText(named: "QuillCodeBrowserCommentState.swift")
         let browserSurfaceText = try Self.appSourceText(named: "QuillCodeBrowserSurface.swift")
         let browserEngineText = try Self.appSourceText(named: "WorkspaceBrowserEngine.swift")
 
-        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserState"), "Browser state should live in a focused surface file.")
-        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserTabState"), "Browser tab state should live in the focused browser surface file.")
-        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserSnapshotState"), "Browser snapshot state should live in a focused surface file.")
-        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserCommentState"), "Browser comment state should live in a focused surface file.")
+        XCTAssertTrue(browserStateText.contains("public struct BrowserState"), "Browser state should live in a focused state file.")
+        XCTAssertTrue(browserTabStateText.contains("public struct BrowserTabState"), "Browser tab state should live in a focused state file.")
+        XCTAssertTrue(browserSnapshotStateText.contains("public struct BrowserSnapshotState"), "Browser snapshot state should live in a focused state file.")
+        XCTAssertTrue(browserCommentStateText.contains("public struct BrowserCommentState"), "Browser comment state should live in a focused state file.")
         XCTAssertTrue(browserSurfaceText.contains("public struct BrowserSurface"), "Browser presentation surface should live in the browser surface file.")
         XCTAssertTrue(browserSurfaceText.contains("public struct BrowserTabSurface"), "Browser tab presentation should live in the browser surface file.")
         XCTAssertTrue(browserSurfaceText.contains("public struct BrowserSnapshotSurface"), "Browser snapshot presentation should live in the browser surface file.")
         XCTAssertTrue(browserSurfaceText.contains("public struct BrowserCommentSurface"), "Browser comment presentation should live in the browser surface file.")
+        XCTAssertFalse(browserSurfaceText.contains("public struct BrowserState"), "Browser presentation surfaces should not own browser reducer state.")
+        XCTAssertFalse(browserStateText.contains("public struct BrowserSurface"), "Browser state should not own browser presentation surfaces.")
         XCTAssertTrue(browserEngineText.contains("BrowserInspector.snapshot"), "Browser state transitions should own browser snapshot construction.")
         XCTAssertTrue(browserEngineText.contains("static func newTab"), "Browser tab creation should be a focused engine transition.")
         XCTAssertTrue(browserEngineText.contains("static func selectTab"), "Browser tab selection should be a focused engine transition.")
